// src/app/core/api/model/platformComponentResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlatformComponentResourceStatusDto } from './platformComponentResourceStatusDto';


export interface PlatformComponentResponseDto { 
    key: string;
    name: string;
    description: string;
    category: string;
    managedBy: PlatformComponentResponseDto.ManagedByEnum;
    status: PlatformComponentResponseDto.StatusEnum;
    restartSupported: boolean;
    errorCount: number;
    errors: Array<string>;
    resources: Array<PlatformComponentResourceStatusDto>;
    checkedAt: string;
}
export namespace PlatformComponentResponseDto {
    export const ManagedByEnum = {
        Flui: 'flui',
        K3s: 'k3s',
        Addon: 'addon'
    } as const;
    export type ManagedByEnum = typeof ManagedByEnum[keyof typeof ManagedByEnum];
    export const StatusEnum = {
        Healthy: 'healthy',
        Degraded: 'degraded',
        Missing: 'missing'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}