// src/app/core/api/model/platformComponentResourceStatusDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlatformComponentConditionDto } from './platformComponentConditionDto';
import { PlatformComponentPodIssueDto } from './platformComponentPodIssueDto';
import { PlatformComponentReplicaStatusDto } from './platformComponentReplicaStatusDto';
import { PlatformComponentPodStatusDto } from './platformComponentPodStatusDto';


export interface PlatformComponentResourceStatusDto { 
    kind: string;
    name: string;
    namespace: string;
    exists: boolean;
    status: PlatformComponentResourceStatusDto.StatusEnum;
    restartSupported: boolean;
    createdAt?: string;
    replicas?: PlatformComponentReplicaStatusDto;
    conditions?: Array<PlatformComponentConditionDto>;
    podIssues?: Array<PlatformComponentPodIssueDto>;
    pods?: Array<PlatformComponentPodStatusDto>;
}
export namespace PlatformComponentResourceStatusDto {
    export const StatusEnum = {
        Healthy: 'healthy',
        Degraded: 'degraded',
        Missing: 'missing'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}