// src/app/core/api/model/redeployPlatformComponentResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RedeployPlatformComponentResponseDto { 
    componentKey: string;
    restartedResources: Array<string>;
    missingResources: Array<string>;
    skippedResources: Array<string>;
    result: RedeployPlatformComponentResponseDto.ResultEnum;
    message: string;
    executedAt: string;
}
export namespace RedeployPlatformComponentResponseDto {
    export const ResultEnum = {
        Ok: 'ok',
        Partial: 'partial',
        Skipped: 'skipped'
    } as const;
    export type ResultEnum = typeof ResultEnum[keyof typeof ResultEnum];
}